+++ documentation/changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (tries to) adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- SOL006 to SOL001 Conversion
+- New flag: Nth List Elem
+
+### Fixed
+- Not keeping empty list as list under certain circumstances
+- Not properly setting boot lists for SOL1 -> SOL6
+
 ## [0.7.0]
 ### Added
 - Unit tests